+++ src/main/java/racingcar/Application.java
@@ -1,8 +1,6 @@
 package racingcar;
 
 import camp.nextstep.edu.missionutils.Console;
-import racingcar.domain.race.RaceController;
-import racingcar.domain.race.RaceService;
 
 public class Application {
     public static void main(String[] args) {
