+++ src/main/java/racingcar/RaceService.java
@@ -1,4 +1,4 @@
-package racingcar.domain.race;
+package racingcar;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.Comparator;
