+++ src/main/java/racingcar/domain/car/CarDtoMapper.java
@@ -2,7 +2,7 @@ package racingcar.domain.car;
 
 import java.util.ArrayList;
 import java.util.List;
-import racingcar.domain.utils.StringTrimmer;
+import racingcar.domain.utils.StringParser;
 
 public final class CarDtoMapper {
     private CarDtoMapper() {
@@ -10,7 +10,7 @@ public final class CarDtoMapper {
 
     public static List<Car> mapToCar(CarDto carDto) {
         String carNamesWithCommas = carDto.carNamesWithCommas();
-        String[] trimAndSplitCarNames = StringTrimmer.trimAndSplit(carNamesWithCommas);
+        String[] trimAndSplitCarNames = StringParser.trimAndSplit(carNamesWithCommas);
 
         List<Car> cars = new ArrayList<>();
         for (String carName : trimAndSplitCarNames) {
