+++ src/test/java/racingcar/RaceServiceTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain.race;
+package racingcar;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
