+++ src/test/java/racingcar/RaceViewTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain.race;
+package racingcar;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
